requirements.txt
mpmath==1.2.1
Pillow==8.4.0
PyYAML==6.0
sympy==1.9
